Use 65534 as the nobody uid
This is part of https://fedoraproject.org/wiki/Changes/RenameNobodyUser.
Essentially, nfsnobody name is dropped, and nobody is used as the name
for the "kernel overflow user". The description field is changes to be
a bit more explanatory. The gid mapping is changed accordingly.

The note about nfsnobody being (uint64_t) -2 is dropped. I don't think
it was at all true, since nfs-utils has a script that always creates
"nfsnobody" with uid and gid of 65534.