lang.{sh,csh}: reset locale to C.UTF-8 if inherited settings are invalid
This operation is done with very low priority: any explicit configuration
in /etc/locale.conf or ~/.i18n is still read.

To check whether the locale settings are invalid, we run locale and check
if it prints anything to stderr. This might seem like an overly broad condition,
but I think in practice it should work well. locale (without any arguments)
only ever warns if the locale is invalid. Trying to figure out which
specific part of the configuration is wrong would be too complicated, because
of the strange inheritance rules between the individual $LC_* variables,
$LC_ALL, and $LANG. Thus, if locale says something is wrong, we simply
unset $LC_ALL and reset the other variables to C.UTF-8.
($LC_ALL is treated a bit different because it is not supposed to be set by
us, so we only unset, and let the other variables play their role.)

This implements
https://fedoraproject.org/wiki/Changes/Reset-locale-if-not-available.