Remove the commas from aliases in the services file
According to the man page for services: https://man7.org/linux/man-pages/man5/services.5.html,
The alias is an optional setting that is a space or tab seperated list. In a previous PR that
changed the services file there was a change introduced that added an alias but had commas
seperating the aliases.
This small change breaks some things downstream, example for this is the Augeas tool breaks
when attempting to read the services file if there is an comma in the alias, I have created
an issue for that here: https://github.com/hercules-team/augeas/issues/772
This PR simply switches out the commas for a space.